pyproject.toml
[tool.poetry]
name = "chirp"
version = "0.1.0"
description = "A bioacoustics research project."
authors = ["Chirp Team <chirp-bio@google.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
# These dependencies are part of the implicit 'main' group, meaning
# they will be installed when using `pip install`.
# Other dependency groups will not be installed by pip, however.
# See: https://python-poetry.org/docs/managing-dependencies/#dependency-groups
python = ">=3.10,<3.12"
perch_hoplite = {git = "https://github.com/google-research/perch-hoplite.git"}
apache-beam = {version = "^2.50.0", extras = ["gcp"]}
chex = "^0.1.7"
clu = "^0.0.12"
jax = "^0.6"
flax = "^0.10"
imageio = "^2.5.0"
ipywidgets = "^8.1"
optax = ">=0.2.3"
matplotlib = "^3.6.1"
notebook = "^7.0.4"
pandas = {version = "^2.1.1", extras=["gcp"]}
numpy = "^1.24"
SPARQLWrapper = "^2.0.0"
tqdm = "^4.6"
tensorflow = "^2.19"
tf-keras = "^2.19"
tensorflow-hub = "^0.16"
wfdb = "^4.1.0"
scikit-learn = "^1.3.0"
google-cloud-storage = "^2.10.0"
joblib = "^1.3.0"

[tool.poetry.group.jaxtrain]
optional = true

[tool.poetry.group.jaxtrain.dependencies]
# TF-io is used for sample rate conversion only in data ingestion pipelines.
# Note that tensorflow and tensorflow-io versions must be synchronized.
# See: https://github.com/tensorflow/io#tensorflow-version-compatibility
tensorflow-io = "^0.37"
tensorflow-datasets = {version = "^4.9", extras = ["dev"]}
# AudioMAE / ViT.
scenic = {git = "https://github.com/google-research/scenic.git"}
# Used for crawling datasources.
ratelimiter = "^1.2.0.post0"
# Quantization-aware training in Jax.
aqtp = "^0.5.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
